Fixes errors with debug torch, /oc groups, now signs will translate on break

+++ src/main/java/ua/mcchickenstudio/opencreative/listeners/player/DestroyBlockListener.java
@@ -80,6 +80,11 @@ public final class DestroyBlockListener implements Listener {
                 event.setCancelled(true);
             }
 
+            if (block.getType() == Material.REDSTONE_WALL_TORCH) {
+                devPlanet.setCodeChanged(true);
+                return;
+            }
+
             if (devPlanet.getAllCodingBlocksForPlacing().contains(block.getType())) {
                 if (ActionCategory.getByMaterial(block.getType()) != null) {
                     platform.destroyCodingBlock(block.getLocation(), devPlanet.isDropItems());
